feat: production manager (#8)
ProductionManager class that manages productions and associated lines and connection information for each line. A client will initiate a connection to an existing production and line only if that production and line exists in the current ProductionManager instance. If this is the case then connection setup will proceed. If not then an error will be raised.
This allows the client to initiate a connection using the name of the production and the name of the line.